requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiohttp==3.9.1
aiorwlock==1.3.0
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.39.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
commonmark==0.9.1
confparser==0.1.7
contourpy==1.2.0
cycler==0.12.1
datasets==2.4.0
decorator==5.1.1
deepspeed==0.9.5
defusedxml==0.7.1
dill==0.3.5.1
einops==0.7.0
et-xmlfile==1.1.0
evaluate==0.2.2
exceptiongroup==1.2.0
fastapi==0.103.1
filelock==3.13.1
Flask==3.0.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h11==0.14.0
hanziconv==0.3.2
hjson==3.1.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.3
jira==3.5.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightgbm==4.1.0
loguru==0.7.0
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.6.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.13
ninja==1.11.1.1
nltk==3.8.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.1
openpyxl==3.1.2
packaging==23.2
pandas==1.2.5
peft==0.4.0
pillow==10.2.0
protobuf==4.23.4
psutil==5.9.8
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==5.3.1
ray==2.6.3
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
retry==0.9.2
rich==12.5.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.13.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.62.3
transformers==4.30.2
transformers-stream-generator==0.0.4
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
xlrd==1.2.0
xxhash==3.4.1
yarl==1.9.4
